backend/requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Database
sqlalchemy[asyncio]>=2.0.0
sqlmodel>=0.0.21
alembic>=1.13.0
asyncpg>=0.30.0
greenlet>=3.0.0

# Authentication
pyjwt>=2.8.0
email-validator>=2.0.0

# HTTP Client (Postmark/Resend)
httpx>=0.27.0

# AI/LLM
langchain>=0.3.0
langchain-openai>=0.2.0
openai>=1.50.0

# CSV Processing
python-multipart>=0.0.9

# Utilities
python-dotenv>=1.0.0
